requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1

# LangChain and AI
langchain==0.1.4
langchain-community==0.0.13
langchain-core==0.1.17
langchain-google-genai==0.0.5
langchain-huggingface==0.0.4
langchain-text-splitters==0.0.1

# Document processing
pypdf==3.17.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# File handling
python-multipart==0.0.6 